Added another fx, a loop, input()
Repurposed initial fx to output the winner of the game. Added in a fx so that the computer will choose between rock, paper, or scissors, for each player. Added a for loop so that the game only runs twice. Added an input() to make it a little more personable for the user.